Refactor: Balance 보단 Cashier 가 더 적절해 클래스명 변경함
